Always raise trapDisarm on probe attempt
Previously, trapDisarm would only be raised when the reference has
attached lockData, and that lockData had a trap.  Now it is raised
every time the player uses a probe.

Adds a new event data item to trapDisarm: trapPresent (bool) - true if
there was a trap present; false otherwise.

Adds a new output event data item to trapDisarm: clearTarget.  When
this is set in the event handler, the player's target is cleared, the
same as when a successful disarm occurs.